Add functions into the modules
