+++ src/index.js
@@ -4,14 +4,16 @@ import { forEach } from 'lodash';
 import form, { todoInput, doList } from './modules/do.js';
 // import SaveTodo from './modules/savetodo.js';
 import Tasks from './modules/tasks.js';
+import { todos,editTodoId } from "./modules/tasks.js";
+
 
 // eslint-disable-line
 
 import './style.css';
 
-// Vars
-export let todos = JSON.parse(localStorage.getItem('todos')) || [];
-export let editTodoId = -1;
+// // Vars
+//  let todos = JSON.parse(localStorage.getItem('todos')) || [];
+//  let editTodoId = -1;
 
 // render todos
 
@@ -31,12 +33,43 @@ export const renderTodos = () => {
                   data-action='check'
                   ></i>
                   <p data-action='edit'>${todo.description}</p>
-                  <i class="fa-solid fa-trash" data-action='delete'><i class="fa-solid fa-ellipsis-vertical" data-action='move'></i>
+                  <i class="fa-solid fa-trash" data-action='delete'>                <i class="fa-solid fa-ellipsis-vertical" data-action='move'></i>
                   </i>
           </div>`;
   });
 };
 
+// // Save todo
+
+// const saveTodo = () => {
+//   const todoValue = todoInput.value;
+
+//   // check if the todolist is empty
+//   const isEmpty = todoValue === '';
+
+//   if (isEmpty) {
+//     alert('To do list is empty');
+//   } else if (todos.some((todo) => todo.description.toUpperCase() === todoValue.toUpperCase())) {
+//     alert('Todo already exists!');
+//   } else {
+//     if (editTodoId >= 0) {
+//       // update the edit todo
+//       todos = todos.map((todo, index) => ({
+//         ...todo,
+//         description: index === editTodoId ? todoValue : todo.description,
+//       }));
+
+//       editTodoId = -1;
+//     } else {
+//       todos.push({
+//         indexNum: todos.length,
+//         description: todoValue,
+//         completed: false,
+//       });
+//     }
+//     todoInput.value = '';
+//   }
+// };
 
 // Form submit
 form.addEventListener('submit', (e) => {
@@ -51,41 +84,40 @@ form.addEventListener('submit', (e) => {
 
 renderTodos();
 
-// check for todo
+// // check for todo
 
-const checkTodo = (itemId) => {
-  todos = todos.map((todo, index) => ({
-    ...todo,
-    completed: index === itemId ? !todo.completed : todo.completed,
-  }));
+// const checkTodo = (itemId) => {
+//   todos = todos.map((todo, index) => ({
+//     ...todo,
+//     completed: index === itemId ? !todo.completed : todo.completed,
+//   }));
 
-  renderTodos();
-  localStorage.setItem('todos', JSON.stringify(todos));
-};
+//   renderTodos();
+//   localStorage.setItem('todos', JSON.stringify(todos));
+// };
 
-// // edit todo
+// //  // edit todo
 
-const editTodo = (itemId) => {
-  todoInput.value = todos[itemId].description;
-  editTodoId = itemId;
-};
+// const editTodo = (itemId) => {
+//   todoInput.value = todos[itemId].description;
+//   editTodoId = itemId;
+// };
 
-// // delete todo
+// // // delete todo
 
-const deleteTodo = (itemId) => {
-  todos = todos.filter((todo, index) => index !== itemId);
-  editTodoId = -1;
+// const deleteTodo = (itemId) => {
+//   todos = todos.filter((todo, index) => index !== itemId);
+//   editTodoId = -1;
 
-  todos.forEach( (todo, index) => {
-    todo.indexNum = index;
-  })
+//   todos.forEach( (todo, index) => {
+//     todo.indexNum = index;
+//   })
 
-  // re render
-  renderTodos();
-  localStorage.setItem('todos', JSON.stringify(todos));
+//   // re render
+//   renderTodos();
+//   localStorage.setItem('todos', JSON.stringify(todos));
 
-  console.log(itemId);
-};
+// };
 
 // add eventlisner for all the do lists
 
@@ -100,8 +132,8 @@ doList.addEventListener('click', (e) => {
 
   const { action } = target.dataset;
 
-  if (action === 'check' && checkTodo(itemId)) return;
-  if (action === 'edit' && editTodo(itemId)) return;
+  if (action === 'check' && Tasks.checkTodo(itemId)) return;
+  if (action === 'edit' && Tasks.editTodo(itemId)) return;
   // action === 'move' && moveTodo(itemId);
-  if (action === 'delete' && deleteTodo(itemId));
+  if (action === 'delete' && Tasks.deleteTodo(itemId));
 });
